+++ app/src/main/java/io/horizontalsystems/ethereumkit/sample/BalanceFragment.kt
@@ -9,7 +9,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
-import io.horizontalsystems.ethereumkit.EthereumKit
+import io.horizontalsystems.ethereumkit.core.EthereumKit
 
 class BalanceFragment : Fragment() {
