+++ app/src/main/java/io/horizontalsystems/ethereumkit/sample/SendReceiveFragment.kt
@@ -10,7 +10,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.*
-import io.horizontalsystems.ethereumkit.models.FeePriority
 
 class SendReceiveFragment : Fragment() {
