+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/api/models/LastBlockHeight.kt
@@ -1,4 +1,4 @@
-package io.horizontalsystems.ethereumkit.models
+package io.horizontalsystems.ethereumkit.api.models
 
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
