+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/core/ERC20.kt
@@ -1,6 +1,6 @@
 package io.horizontalsystems.ethereumkit.core
 
-import io.horizontalsystems.ethereumkit.EthereumKit
+import io.horizontalsystems.ethereumkit.core.EthereumKit
 
 class ERC20(var contractAddress: String, var listener: EthereumKit.Listener) {
