+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/network/Network.kt
@@ -1,3 +1,3 @@
-package io.horizontalsystems.ethereumkit.models
+package io.horizontalsystems.ethereumkit.network
 
 enum class NetworkType { MainNet, Ropsten, Kovan, Rinkeby }
